claas/requirements.txt
coverage==5.5
pytest==6.2.2
pytest-cov==2.11.1
pylint==2.7.4
PyYAML==5.3.1
mock==4.0.3